FIX BIANYI

+++ pkg/src/cluster/dtc_recovery.c
@@ -1680,7 +1680,7 @@ void dtc_standby_update_lrp(knl_session_t *session, uint32 idx, uint32 size_read
     }
 
     dtc_node_ctrl_t *ctrl = dtc_get_ctrl(session, idx);
-    CT_LOG_RUN_INF("[LCM DEBUG] ctrl lsn %llu lfn %llu ,log end lsn %llu, lfn %llu", ctrl->lsn, ctrl->lfn, tail->point.lsn, tail->point.lfn);
+    CT_LOG_RUN_INF("[LCM DEBUG] ctrl lsn %llu lfn %llu ,log end lsn %llu, lfn %llu", ctrl->lsn, ctrl->lfn, tail->point.lsn, (uint64)tail->point.lfn);
     if (ctrl->lrp_point.lsn < tail->point.lsn) {
         ctrl->lrp_point = tail->point;
         ctrl->scn = DB_CURR_SCN(session);
